refactor(perfmodel): move parse_bool helper to utils
Relocate parse_bool from perf_model.py into PerfModel.utils per PR review.
